Control SnowSQL install location.
By default, when SnowSQL updates, it puts the files it needs in `~/.snowsql`.
That doesn't work well for Sqitch, because we'll often mount the home
directory to the originating host. SnowSQL then would not find the files it
needs, unless they happened to be installed on the originating host. Even then
they might not be compatible.

So we use the (currently undocumented) `$WORKSAPCE` environment variable to
tell SnowSQL where to install and find its files. While at it, overwrite the
configuration file there with the account name and to disable auto-update.