Build Snowflake prereqs in separate container.
Then copy them over. That way we don't end up with earlier layers with extra
crap in them.

And to better support building the Snowflake image, which requires a build
arg, specify the directory and registry prefix via environment variales
instead of arguments, then pass any and all arguments to `docker build`.
That lets one build the Snowflake image and tag it for a private registry
like so:

    env DIR=snowflake REGISTRY=registry.example.com/sqitch \
    ./build --build-arg sf_account=$ACCOUNT