Change home directory mountpoint to /root
The PostgreSQL libraries do not honor the `$HOME` environment variable
when looking for configuration files like `.pgpass`.  This change amends
the Dockerfile to mount the user's home directory in `/root` to conform
to the `/etc/passwd` entry for the root user.  This allows `sqitch` to
successfully locate `.pgpass` for credentials when connecting to
`db:pg:` targets.

This could alternatively be solved by overriding or changing
`/etc/passwd` to point to `/home` for the root user, but this approach
seemed the least invasive.