Fix SnowSQL permission issues
Turns out that SnowSQL uses the WORKSPACE variable to find both its binary
downloads *and* it's configuration file. So Move the snowsql app to /var and
make /bin/snowsql a shell script that always passes `--config` to point to the
mouted configuration file.

Futhermore, SnowSQL wants to write to files in its home directory
subdirectory, and not just the config file. Maybe log files? So make the
sqitch user the owner of those files so that SnowSQL can write stuff without
error.