Run with host uid and gid on linux.
Rather than just connecting as root when running as root, on Linux, always
connect with the host uid and gid, so that when the directories are mounted,
they can be written to from inside the container. But don't do it on macOS
(darwin), where the container runs inside a VM and does not have the Mac
uid and gid owning the files, because the VM user owns them.

Done by moving the OS-specific logic into one place after the `passopt`
variable is initialized.

Resolves #8.