Install CA certs.
And create a symlink so that libcurl can find them. Discovered when attempting
to access Snowflake via a web proxy, but may well apply to other engines, so
including here in the core image. Thanks to Egan Neuhengen (with some input
from me and from Snowflake support) for diagnosing the problem and figuring
out the solution.