Remove snowdql shell script hack.
It worked around a lack of distinction between the location of thew SnowSQL
binaries and config files, but didn't work when the containuer was run as
root, becase it hard-coded the config file location to `/home`, but for root
the home directory is `/root`. But thanks to the introduction of the
`SNOWSQL_DOWNLOAD_DIR` environment variable in b02f8bb, we no longer need
the hack. So restore the SnowSQL binary to `/bin` and remove the shell
script (reversing 916aab7 and 81921f0).