Build images for ARM64
Update the CI/CD workflow to use the latest Docker actions for
assembling metadata, building, testing, and pushing multi-platform
Docker images. Cease using the `./build` script in the workflow, but
update it to build images for multiple platforms. To have both the CI/CD
workflow and the `./build` script in sync, add the `.envrc` file and
teach both to read from it for common values, especially the Sqitch
version.

Fix an issue building Sqitch on ARM64 by installing List::MoreUtils::XS
before building Sqitch; otherwise it hangs and times out during the
configure phase for that module.

Update the Oracle `Dockerfile` to always install the latest Instant
Client; no more need to edit it for new versions. Also teach it to
properly install ARM64 client libraries, but disable building for ARM64
in the `Makefile`, for now, until a [bug building DBD::Oracle] can be
fixed.

Update the Snowflake `Dockerfile` to install the ARM64 ODBC driver on
ARM64 Linux, but limit builds to AMD64 in the `Makefile`, for now, until
[SnowSQL gets ARM64 support].

Limit the Exasol build to AMD64, as there are currently no Exasol
libraries for ARM64.

Remove reference to Vertica, for which there is not yet Docker support.

  [bug building DBD::Oracle]: https://rt.cpan.org/Ticket/Display.html?id=149876
  [SnowSQL gets ARM64 support]: https://community.snowflake.com/s/question/0D5Do00000ltxpVKAQ/snowsql-linuxarm64-support