Fix Snowflake ODBC configuration
The `SF_ACCOUNT` item under `ODBC Data Sources` was incorrectly pointing
at a description instead of driver named in `odbcinst.ini`. As a result
none of its configs ever did anything. Fix it to properly point to
`Snowflake`, the default driver named in `odbcinst.ini` and keep only
the config necessary for it to operate, `Driver`. But also add a comment
listing Snowflake key pair configs.

Speaking of the `odbcinst.ini` entry, fix its `Driver` config to point
to the actual location of the `libSnowflake.so` file. With these
changes, the resulting Docker image properly works with Snowflake.

While at it, fix shell linting issues in `docker-sqitch.sh` and add some
notes about custom configuration to the Snowflake README.