Fix ClickHouse tags
They will now be this style:

  sqitch/sqitch:latest-clickhouse
  sqitch/sqitch:clickhouse
  sqitch/sqitch:v1.6-clickhouse
  sqitch/sqitch:v1.6.0-clickhouse
  sqitch/sqitch:v1.6.0.1-clickhouse

While at it, prevent build failure when testing before release, when a
new version won't be on Docker Hub yet, by setting the version to
"latest" when building the image.

Cease building the ClickHouse image on arm64, as the ODBC driver is
amd64-only for now. Update the README to reflect the publication of
ClickHouse-supporting images, and the lack of arm64 support for
ClickHouse.